refactor: group 쿼리 위치이동
